스터디 리턴 변경 및 스웨거 설명 코드 추가

+++ backend/sooksook/src/main/java/com/smwuis/sooksook/service/study/StudyMemberService.java
@@ -6,7 +6,8 @@ import com.smwuis.sooksook.domain.study.StudyMember;
 import com.smwuis.sooksook.domain.study.StudyMemberRepository;
 import com.smwuis.sooksook.domain.user.User;
 import com.smwuis.sooksook.domain.user.UserRepository;
-import com.smwuis.sooksook.web.dto.study.StudyMemberListResponseDto;
+import com.smwuis.sooksook.web.dto.study.StudyBoardResponseDto;
+import com.smwuis.sooksook.web.dto.study.StudyMemberResponseDto;
 import com.smwuis.sooksook.web.dto.study.StudyMemberSaveRequestDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -31,11 +32,9 @@ public class StudyMemberService {
         StudyBoard studyBoard = studyBoardRepository.findById(boardId).orElseThrow(()-> new IllegalArgumentException("해당 게시판이 없습니다."));
 
         // 가입하지 않은 멤버라면
-        if (studyMemberRepository.findByStudyBoardIdAndUserId(studyBoard, user) == null) {
-
+        if (!studyMemberRepository.findByStudyBoardIdAndUserId(studyBoard, user).isPresent()) {
             return false;
         }
-
         // 가입한 멤버라면
         else {
             return true;
@@ -44,7 +43,7 @@ public class StudyMemberService {
 
     // 스터디 게시판 비밀번호 확인 및 스터디 가입
     @Transactional
-    public Boolean password(StudyMemberSaveRequestDto saveRequestDto) {
+    public StudyMemberResponseDto password(StudyMemberSaveRequestDto saveRequestDto) {
         StudyBoard studyBoard = studyBoardRepository.findById(saveRequestDto.getStudyBoardId()).orElseThrow(()-> new IllegalArgumentException("해당 게시판이 없습니다."));
 
         if(studyBoard.getPassword().equals(saveRequestDto.getPassword())) {
@@ -53,45 +52,55 @@ public class StudyMemberService {
             joinMember.setUser(userRepository.findByEmail(saveRequestDto.getEmail()).orElseThrow(()-> new IllegalArgumentException("해당 유저가 없습니다.")));
             studyBoard.addStudyMember(studyMemberRepository.save(joinMember));
             studyMemberRepository.save(joinMember);
-            return true;
+            return new StudyMemberResponseDto(joinMember);
         }
-
         else {
-            return false;
+            throw new RuntimeException("스터디 게시판 가입에 실패했습니다.");
         }
     }
     
     // 스터디 탈퇴
     @Transactional
-    public Long drop(String email, Long studyBoardId) {
+    public Boolean drop(String email, Long studyBoardId) {
         StudyBoard studyBoard = studyBoardRepository.findById(studyBoardId).orElseThrow(()-> new IllegalArgumentException("해당 게시판이 없습니다."));
         User user = userRepository.findByEmail(email).orElseThrow(()-> new IllegalArgumentException("해당 유저가 없습니다."));
-        StudyMember studyMember = studyMemberRepository.deleteByStudyBoardIdAndUserId(studyBoard, user);
-        return user.getId();
+        
+        if(user.getEmail().equals(email)) {
+            StudyMember studyMember = studyMemberRepository.deleteByStudyBoardIdAndUserId(studyBoard, user);
+            return true;
+        }
+        else {
+            throw new RuntimeException("스터디 게시판 탈퇴에 실패했습니다.");
+        }
     }
 
     // 스터디 별 정보 조회 (멤버 이름, 글 작성 수와 댓글 작성 수 등)
     @Transactional(readOnly = true)
-    public List<StudyMemberListResponseDto> findByAllByStudyBoardId(Long studyBoardId) {
+    public List<StudyMemberResponseDto> findByAllByStudyBoardId(Long studyBoardId) {
         StudyBoard studyBoard = studyBoardRepository.findById(studyBoardId).orElseThrow(()-> new IllegalArgumentException("해당 게시판이 없습니다."));
         return studyMemberRepository.findAllByStudyBoardId(studyBoard)
                 .stream()
-                .map(StudyMemberListResponseDto::new)
+                .map(StudyMemberResponseDto::new)
                 .collect(Collectors.toList());
     }
 
     // 내가 참여 중인 스터디
     @Transactional(readOnly = true)
-    public List<Long> myStudy(String email) {
+    public List<StudyBoardResponseDto> myStudy(String email) {
         User user = userRepository.findByEmail(email).orElseThrow(()-> new IllegalArgumentException("해당 유저가 없습니다."));
         List<StudyMember> studyList = studyMemberRepository.findAllByUserId(user);
         List<Long> studyBoardIdList = new ArrayList<>();
+        List<StudyBoard> studyBoardResponseDtoList = new ArrayList<>();
 
         for(StudyMember studyMember: studyList) {
             studyBoardIdList.add(studyMember.getStudyBoardId().getId());
         }
 
-        return studyBoardIdList;
+        for(Long studyBoardId: studyBoardIdList) {
+            studyBoardResponseDtoList.add(studyBoardRepository.findById(studyBoardId).orElseThrow(()-> new IllegalArgumentException("해당 게시판이 없습니다.")));
+        }
+
+        return studyBoardResponseDtoList.stream().map(StudyBoardResponseDto::new).collect(Collectors.toList());
     }
 }
